Refactored Room Popups
